backend/requirements.txt
django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
djoser==2.1.0
markdown==3.4.1
django-filter==22.1
psycopg2==2.9.5
pandas==1.5.2